fix fal inputs
