+++ src/components/Window/WindowMenu/WindowMenu.tsx
@@ -7,7 +7,7 @@ import { nanoid } from 'nanoid'
 export const WindowMenu: React.FC<{
   id: string
 }> = ({ id }) => {
-  const state = useStore(['addContentToItem'])
+  const state = useStore([])
   return (
     <div className={style.wrapper}>
       <Dropdown.Menu
@@ -15,29 +15,9 @@ export const WindowMenu: React.FC<{
         SelectedOption={() => <p>New</p>}
         Options={[
           <Dropdown.Item
-            onClick={() => {
-              state.addContentToItem(id, {
-                content: 'New Text',
-                type: 'text',
-                id: nanoid(),
-              })
-            }}
-            key={'Text'}
-            label1="Text Block"
-            isChecked={false}
-          />,
-          <Dropdown.Item
-            onClick={() => {
-              state.addContentToItem(id, {
-                type: 'canvas',
-                content: {
-                  base64: '',
-                },
-                id: nanoid(),
-              })
-            }}
-            key={'Canvas'}
-            label1="Canvas"
+            onClick={() => {}}
+            key={'Pin'}
+            label1="Pin"
             isChecked={false}
           />,
         ]}
