+++ src/mock/mockNouns.ts
@@ -3,26 +3,16 @@ export const MOCK_NOUNS = [
   'banana',
   'cat',
   'dog',
-  'elephant',
-  'ferris wheel',
-  'giraffe',
-  'hamburger',
-  'ice cream',
   'jellyfish',
-  'kangaroo',
-  'lion',
   'mountain',
   'notebook',
   'octopus',
-  'piano',
   'quilt',
   'robot',
   'sunflower',
-  'tiger',
   'umbrella',
   'volcano',
   'waterfall',
-  'xylophone',
   'yacht',
   'zebra',
   'airplane',
@@ -32,28 +22,14 @@ export const MOCK_NOUNS = [
   'eagle',
   'fireplace',
   'gorilla',
-  'hat',
   'igloo',
-  'jacket',
-  'kite',
-  'lamp',
-  'microscope',
-  'orange',
-  'panda',
   'queen',
   'rainbow',
-  'sandwich',
   'train',
-  'unicorn',
-  'violin',
-  'whale',
-  'yogurt',
-  'zeppelin',
   'amethyst',
   'butterfly',
   'castle',
   'drum',
-  'earring',
   'fountain',
   'globe',
   'house',
@@ -63,8 +39,6 @@ export const MOCK_NOUNS = [
   'ladder',
   'mirror',
   'nest',
-  'owl',
-  'penguin',
   'quartz',
   'rose',
   'suitcase',
@@ -76,24 +50,13 @@ export const MOCK_NOUNS = [
   'beetle',
   'car',
   'daisy',
-  'eel',
   'flag',
   'guitar',
-  'helicopter',
   'iceberg',
-  'jelly',
   'koala',
   'lemon',
-  'monkey',
-  'nail',
-  'parrot',
-  'quiver',
-  'snake',
-  'tractor',
   'unicycle',
   'vulture',
-  'wagon',
-  'yoyo',
   'anchor',
   'book',
   'cactus',
@@ -113,11 +76,7 @@ export const MOCK_NOUNS = [
   'arrow',
   'bottle',
   'carrot',
-  'fish',
-  'needle',
-  'rabbit',
   'star',
-  'whistle',
   'angel',
   'bridge',
   'cookie',
@@ -171,22 +130,19 @@ export const MOCK_NOUNS = [
   'arch',
   'ball',
   'duck',
-  'firework',
+  'fireworks',
   'juice',
-  'map',
   'otter',
   'peacock',
   'squirrel',
   'armadillo',
   'cherry',
-  'light',
   'olive',
   'trumpet',
   'avocado',
   'caterpillar',
   'feather',
   'river',
-  'aeroplane',
   'chocolate',
   'acorn',
   'boat',
@@ -195,104 +151,59 @@ export const MOCK_NOUNS = [
   'almond',
   'antelope',
   'artichoke',
-  'Bacon',
-  'Ball',
-  'Bike',
-  'Car',
-  'Chair',
-  'Cheese',
-  'Chicken',
-  'Chips',
-  'Computer',
-  'Fish',
-  'Gloves',
-  'Hat',
-  'Keyboard',
-  'Mouse',
-  'Pants',
-  'Pizza',
-  'Salad',
-  'Sausages',
-  'Shirt',
-  'Shoes',
-  'Soap',
-  'Table',
-  'Towels',
-  'Tuna',
-  'Cargo Van',
-  'Convertible',
-  'Coupe',
-  'Crew Cab Pickup',
-  'Extended Hatchback',
-  'Minivan',
-  'Passenger SUV',
-  'Sedan',
-  'Wagon',
-  'ants',
-  'bats',
-  'bears',
-  'bees',
-  'birds',
+  'bacon',
+  'cheese',
+  'mouse',
+  'pizza',
+  'ant',
+  'bat',
+  'bear',
+  'bee',
+  'bird',
   'buffalo',
-  'cats',
-  'chickens',
+  'cat',
+  'chicken',
   'cattle',
-  'dogs',
-  'dolphins',
-  'ducks',
-  'elephants',
-  'fishes',
-  'foxes',
-  'frogs',
+  'dog',
+  'dolphin',
+  'duck',
+  'elephant',
+  'fish',
+  'fox',
+  'frog',
   'geese',
-  'goats',
-  'horses',
-  'kangaroos',
-  'lions',
-  'monkeys',
-  'owls',
-  'oxen',
-  'penguins',
-  'people',
-  'pigs',
-  'rabbits',
+  'goat',
+  'horse',
+  'kangaroo',
+  'lion',
+  'monkey',
+  'owl',
+  'penguin',
+  'pig',
+  'rabbit',
   'sheep',
-  'tigers',
-  'whales',
-  'wolves',
-  'zebras',
+  'tiger',
+  'whale',
+  'wolf',
+  'zebra',
   'banshees',
   'crows',
   'black cats',
-  'chimeras',
   'ghosts',
   'conspirators',
   'dragons',
   'dwarves',
   'elves',
-  'enchanters',
-  'exorcists',
   'sons',
   'foes',
-  'giants',
-  'gnomes',
-  'goblins',
-  'gooses',
-  'griffins',
-  'lycanthropes',
-  'nemesis',
-  'ogres',
-  'oracles',
-  'prophets',
-  'sorcerors',
-  'spiders',
-  'spirits',
-  'vampires',
-  'warlocks',
-  'vixens',
-  'werewolves',
-  'witches',
-  'worshipers',
-  'zombies',
-  'druids',
+  'giant',
+  'gnome',
+  'goblin',
+  'ogre',
+  'spider',
+  'spirit',
+  'vampire',
+  'werewolf',
+  'witch',
+  'zombie',
 ]
