+++ src/ui/Dropdown.module.scss
@@ -33,7 +33,25 @@
   }
 }
 
+@mixin line {
+  & + .item,
+  & + .subMenu {
+    &::before {
+      background-color: var(--white-15);
+      opacity: 0.5;
+      position: absolute;
+      top: 0;
+      left: 0;
+      content: '';
+      width: 90%;
+      transform: translateX(5%);
+      height: 1px;
+    }
+  }
+}
+
 .subMenu {
+  @include line;
   user-select: none;
   display: flex;
   gap: 10px;
@@ -54,6 +72,20 @@
   border-radius: 5px;
   &:hover {
     background-color: var(--white-15);
+  }
+  .subMenuList {
+    left: calc(100% + 10px);
+    position: absolute;
+    top: 0;
+    &::before {
+      content: '';
+      position: absolute;
+      top: 0;
+      left: -10px;
+      width: 10px;
+      height: 100%;
+      opacity: 0.5;
+    }
   }
 }
 
@@ -88,6 +120,7 @@
 }
 
 .item {
+  @include line;
   padding: 10px 10px;
   position: relative;
   display: flex;
@@ -114,21 +147,6 @@
     outline: none;
     border: none;
     border: 1px solid var(--border);
-  }
-  & + .item,
-  & + .subMenu {
-    &::before {
-      background-color: var(--white-15);
-      // @include mixins.divider(var(--white-15));
-      opacity: 0.5;
-      position: absolute;
-      top: 0;
-      left: 0;
-      content: '';
-      width: 90%;
-      transform: translateX(5%);
-      height: 1px;
-    }
   }
   &:hover {
     background-color: var(--white-15);
